pyproject.toml
[tool.poetry]
name = "temp_sensor"
version = "0.1.0"
description = ""
authors = ["Richard Campen <richard@campen.co>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
